# app/controllers.go
// Code generated by goagen v1.3.1, DO NOT EDIT.
//
// API "Model Provider Client": Application Controllers
//
// Command:
// $ goagen
// --design=ModelProvider/design
// --out=$(GOPATH)\src\ModelProvider
// --version=v1.3.1

package app

import (
	"context"
	"github.com/goadesign/goa"
	"github.com/goadesign/goa/cors"
	"net/http"
)

// initService sets up the service encoders, decoders and mux.
func initService(service *goa.Service) {
	// Setup encoders and decoders
	service.Encoder.Register(goa.NewJSONEncoder, "application/json")
	service.Encoder.Register(goa.NewGobEncoder, "application/gob", "application/x-gob")
	service.Encoder.Register(goa.NewXMLEncoder, "application/xml")
	service.Decoder.Register(goa.NewJSONDecoder, "application/json")
	service.Decoder.Register(goa.NewGobDecoder, "application/gob", "application/x-gob")
	service.Decoder.Register(goa.NewXMLDecoder, "application/xml")

	// Setup default encoder and decoder
	service.Encoder.Register(goa.NewJSONEncoder, "*/*")
	service.Decoder.Register(goa.NewJSONDecoder, "*/*")
}

// ModelProviderController is the controller interface for the ModelProvider actions.
type ModelProviderController interface {
	goa.Muxer
	AskData(*AskDataModelProviderContext) error
	Create(*CreateModelProviderContext) error
	Upload(*UploadModelProviderContext) error
	UploadFinal(*UploadFinalModelProviderContext) error
}

// MountModelProviderController "mounts" a ModelProvider resource controller on the given service.
func MountModelProviderController(service *goa.Service, ctrl ModelProviderController) {
	initService(service)
	var h goa.Handler

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewAskDataModelProviderContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.AskData(rctx)
	}
	service.Mux.Handle("POST", "/model/askData/:data_hash/:ETH_key/:contract_hash", ctrl.MuxHandler("askData", h, nil))
	service.LogInfo("mount", "ctrl", "ModelProvider", "action", "AskData", "route", "POST /model/askData/:data_hash/:ETH_key/:contract_hash")

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewCreateModelProviderContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.Create(rctx)
	}
	service.Mux.Handle("POST", "/model/create/:ETH_key/:smart_contract", ctrl.MuxHandler("create", h, nil))
	service.LogInfo("mount", "ctrl", "ModelProvider", "action", "Create", "route", "POST /model/create/:ETH_key/:smart_contract")

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewUploadModelProviderContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.Upload(rctx)
	}
	service.Mux.Handle("POST", "/model/upload/:model_hash/:ETH_key/:HE_key/:RSA_key/:contract_hash", ctrl.MuxHandler("upload", h, nil))
	service.LogInfo("mount", "ctrl", "ModelProvider", "action", "Upload", "route", "POST /model/upload/:model_hash/:ETH_key/:HE_key/:RSA_key/:contract_hash")

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewUploadFinalModelProviderContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.UploadFinal(rctx)
	}
	service.Mux.Handle("POST", "/model/uploadFinal/:hash/:rsa_key/:trans_id/:ETH_key", ctrl.MuxHandler("uploadFinal", h, nil))
	service.LogInfo("mount", "ctrl", "ModelProvider", "action", "UploadFinal", "route", "POST /model/uploadFinal/:hash/:rsa_key/:trans_id/:ETH_key")
}

// SwaggerController is the controller interface for the Swagger actions.
type SwaggerController interface {
	goa.Muxer
	goa.FileServer
}

// MountSwaggerController "mounts" a Swagger resource controller on the given service.
func MountSwaggerController(service *goa.Service, ctrl SwaggerController) {
	initService(service)
	var h goa.Handler
	service.Mux.Handle("OPTIONS", "/swagger.json", ctrl.MuxHandler("preflight", handleSwaggerOrigin(cors.HandlePreflight()), nil))

	h = ctrl.FileHandler("/swagger.json", "swagger/swagger.json")
	h = handleSwaggerOrigin(h)
	service.Mux.Handle("GET", "/swagger.json", ctrl.MuxHandler("serve", h, nil))
	service.LogInfo("mount", "ctrl", "Swagger", "files", "swagger/swagger.json", "route", "GET /swagger.json")
}

// handleSwaggerOrigin applies the CORS response headers corresponding to the origin.
func handleSwaggerOrigin(h goa.Handler) goa.Handler {

	return func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		origin := req.Header.Get("Origin")
		if origin == "" {
			// Not a CORS request
			return h(ctx, rw, req)
		}
		if cors.MatchOrigin(origin, "*") {
			ctx = goa.WithLogContext(ctx, "origin", origin)
			rw.Header().Set("Access-Control-Allow-Origin", origin)
			rw.Header().Set("Access-Control-Allow-Credentials", "false")
			if acrm := req.Header.Get("Access-Control-Request-Method"); acrm != "" {
				// We are handling a preflight request
				rw.Header().Set("Access-Control-Allow-Methods", "GET")
			}
			return h(ctx, rw, req)
		}

		return h(ctx, rw, req)
	}
}

// SwaggerUIController is the controller interface for the SwaggerUI actions.
type SwaggerUIController interface {
	goa.Muxer
	goa.FileServer
}

// MountSwaggerUIController "mounts" a SwaggerUI resource controller on the given service.
func MountSwaggerUIController(service *goa.Service, ctrl SwaggerUIController) {
	initService(service)
	var h goa.Handler

	h = ctrl.FileHandler("/swagger-ui/*filepath", "swagger-ui/")
	service.Mux.Handle("GET", "/swagger-ui/*filepath", ctrl.MuxHandler("serve", h, nil))
	service.LogInfo("mount", "ctrl", "SwaggerUI", "files", "swagger-ui/", "route", "GET /swagger-ui/*filepath")

	h = ctrl.FileHandler("/swagger-ui/", "swagger-ui\\index.html")
	service.Mux.Handle("GET", "/swagger-ui/", ctrl.MuxHandler("serve", h, nil))
	service.LogInfo("mount", "ctrl", "SwaggerUI", "files", "swagger-ui\\index.html", "route", "GET /swagger-ui/")
}