# swagger/swagger.json
{"swagger":"2.0","info":{"title":"Model Provider Client","description":"Model Provider Client","version":""},"host":"localhost:2626","schemes":["http"],"consumes":["application/json","application/xml","application/gob","application/x-gob"],"produces":["application/json","application/xml","application/gob","application/x-gob"],"paths":{"/model/askData/{data_hash}/{ETH_key}/{contract_hash}":{"post":{"tags":["ModelProvider"],"summary":"askData ModelProvider","description":"ask for data","operationId":"ModelProvider#askData","produces":["application/vnd.goa.error","plain/text"],"parameters":[{"name":"ETH_key","in":"path","description":"ETH private key for transaction","required":true,"type":"string"},{"name":"contract_hash","in":"path","description":"smart contract hash","required":true,"type":"string"},{"name":"data_hash","in":"path","description":"data hash","required":true,"type":"string"}],"responses":{"200":{"description":"OK"},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/error"}},"501":{"description":"Not Implemented","schema":{"$ref":"#/definitions/error"}}},"schemes":["http"]}},"/model/create/{ETH_key}/{smart_contract}":{"post":{"tags":["ModelProvider"],"summary":"create ModelProvider","description":"create smart contract","operationId":"ModelProvider#create","produces":["application/vnd.goa.error","plain/text"],"parameters":[{"name":"ETH_key","in":"path","description":"ETH private key for transaction","required":true,"type":"string"},{"name":"smart_contract","in":"path","description":"smart contract","required":true,"type":"string"}],"responses":{"200":{"description":"OK"},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/error"}},"501":{"description":"Not Implemented","schema":{"$ref":"#/definitions/error"}}},"schemes":["http"]}},"/model/upload/{model_hash}/{ETH_key}/{HE_key}/{RSA_key}/{contract_hash}":{"post":{"tags":["ModelProvider"],"summary":"upload ModelProvider","description":"upload model","operationId":"ModelProvider#upload","produces":["application/vnd.goa.error","plain/text"],"parameters":[{"name":"ETH_key","in":"path","description":"ETH private key for transaction","required":true,"type":"string"},{"name":"HE_key","in":"path","description":"Homomorphic Encryption Key","required":true,"type":"string"},{"name":"RSA_key","in":"path","description":"RSA public key","required":true,"type":"string"},{"name":"contract_hash","in":"path","description":"smart contract hash","required":true,"type":"string"},{"name":"model_hash","in":"path","description":"encrypted model IPFS address","required":true,"type":"string"}],"responses":{"200":{"description":"OK"},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/error"}},"501":{"description":"Not Implemented","schema":{"$ref":"#/definitions/error"}}},"schemes":["http"]}},"/model/uploadFinal/{hash}/{rsa_key}/{trans_id}/{ETH_key}":{"post":{"tags":["ModelProvider"],"summary":"uploadFinal ModelProvider","description":"upload final argument and final distance for transaction ID","operationId":"ModelProvider#uploadFinal","produces":["application/vnd.goa.error","plain/text"],"parameters":[{"name":"ETH_key","in":"path","description":"ETH private key for transaction","required":true,"type":"string"},{"name":"hash","in":"path","description":"final arguments hash","required":true,"type":"string"},{"name":"rsa_key","in":"path","description":"rsa private key for transaction id","required":true,"type":"string"},{"name":"trans_id","in":"path","description":"transaction_id","required":true,"type":"integer"}],"responses":{"200":{"description":"OK"},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/error"}},"501":{"description":"Not Implemented","schema":{"$ref":"#/definitions/error"}}},"schemes":["http"]}},"/swagger-ui/{filepath}":{"get":{"summary":"Download swagger-ui/","operationId":"swagger-ui#/swagger-ui/*filepath","parameters":[{"name":"filepath","in":"path","description":"Relative file path","required":true,"type":"string"}],"responses":{"200":{"description":"File downloaded","schema":{"type":"file"}},"404":{"description":"File not found","schema":{"$ref":"#/definitions/error"}}},"schemes":["http"]}},"/swagger.json":{"get":{"summary":"Download swagger/swagger.json","operationId":"swagger#/swagger.json","responses":{"200":{"description":"File downloaded","schema":{"type":"file"}}},"schemes":["http"]}}},"definitions":{"error":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"code":{"type":"string","description":"an application-specific error code, expressed as a string value.","example":"invalid_value"},"detail":{"type":"string","description":"a human-readable explanation specific to this occurrence of the problem.","example":"Value of ID must be an integer"},"id":{"type":"string","description":"a unique identifier for this particular occurrence of the problem.","example":"3F1FKVRR"},"meta":{"type":"object","description":"a meta object containing non-standard meta-information about the error.","example":{"timestamp":1458609066},"additionalProperties":true},"status":{"type":"string","description":"the HTTP status code applicable to this problem, expressed as a string value.","example":"400"}},"description":"Error response media type (default view)","example":{"code":"invalid_value","detail":"Value of ID must be an integer","id":"3F1FKVRR","meta":{"timestamp":1458609066},"status":"400"}}}}